Map class updated.
Map constructor was already created to create a blank map. Underneath this within the constructor I have added code to read the specified file name “world.txt” (don’t think we need to ask user for a file name). The method I have used is to read each line and split each into a String array, split by delimiter “/”, and capture each line in a String array ArrayList so I can access each one sequentially. If the txt file has too many lines (more than 25) then an error is thrown. I have then created a loadMap method which creates areas of the map and stores in the private variable ‘grid’ of the Map class. If any invalid values have been provided an error is thrown. NEED TO CHECK IF THE METHOD WORKS WITH THE ACTOR CLASS.